Refactor Zuora type imports to use barrel files and mapped file paths
