+++ handlers/discount-api/src/eligibilityChecker.ts
@@ -1,12 +1,12 @@
 import { ValidationError } from '@modules/errors';
 import type { Logger } from '@modules/logger';
 import { getIfDefined } from '@modules/nullAndUndefined';
-import type { SimpleInvoiceItem } from '@modules/zuora/billingPreview';
 import { getNextInvoiceTotal } from '@modules/zuora/billingPreview';
+import type { SimpleInvoiceItem } from '@modules/zuora/billingPreview';
+import type { ZuoraSubscription } from '@modules/zuora/types';
 import type { ZuoraCatalogHelper } from '@modules/zuora-catalog/zuoraCatalog';
 import type { Dayjs } from 'dayjs';
 import dayjs from 'dayjs';
-import type { ZuoraSubscription } from '../../../modules/zuora/src/types/objects/subscription';
 
 export class EligibilityChecker {
 	constructor(private logger: Logger) {}
