+++ handlers/discount-expiry-notifier/src/handlers/getOldPaymentAmount.ts
@@ -1,10 +1,10 @@
 import { stageFromEnvironment } from '@modules/stage';
 import { getBillingPreview } from '@modules/zuora/billingPreview';
 import { doQuery } from '@modules/zuora/query';
+import type { BillingPreviewInvoiceItem } from '@modules/zuora/types';
 import { ZuoraClient } from '@modules/zuora/zuoraClient';
 import dayjs from 'dayjs';
 import { z } from 'zod';
-import type { BillingPreviewInvoiceItem } from '../../../../modules/zuora/src/types/objects/billingPreview';
 import { calculateTotalAmount, filterRecords } from '../helpers';
 import {
 	BaseRecordForEmailSendSchema,
