+++ handlers/negative-invoices-processor/src/handlers/detectFailures.ts
@@ -1,7 +1,7 @@
 /* eslint-disable @typescript-eslint/require-await -- this is required to ensure the lambda returns a value*/
 import { DetectFailuresInputSchema } from '../types';
 import type { DetectFailuresInput } from '../types';
-import type { ProcessedInvoice } from '../types/shared';
+import type { ProcessedInvoice } from '../types';
 
 export const handler = async (event: DetectFailuresInput) => {
 	try {
