+++ handlers/product-switch-api/src/switchInformation.ts
@@ -10,11 +10,8 @@ import type { Stage } from '@modules/stage';
 import type { SimpleInvoiceItem } from '@modules/zuora/billingPreview';
 import type { Dayjs } from 'dayjs';
 import dayjs from 'dayjs';
-import type { ZuoraAccount } from '../../../modules/zuora/src/types/objects/account';
-import type {
-	RatePlan,
-	ZuoraSubscription,
-} from '../../../modules/zuora/src/types/objects/subscription';
+import type { ZuoraAccount } from '@modules/zuora/types';
+import type { RatePlan, ZuoraSubscription } from '@modules/zuora/types';
 import type { CatalogInformation } from './catalogInformation';
 import { getCatalogInformation } from './catalogInformation';
 import type { Discount } from './discounts';
