+++ handlers/product-switch-api/test/contributionToSupporterPlusIntegration.test.ts
@@ -16,9 +16,9 @@ import {
 import { createPayment } from '@modules/zuora/payment';
 import { getSubscription } from '@modules/zuora/subscription';
 import { ZuoraClient } from '@modules/zuora/zuoraClient';
+import type { ZuoraSubscription } from '@modules/zuora/types';
+import { zuoraDateFormat } from '@modules/zuora/utils';
 import dayjs from 'dayjs';
-import type { ZuoraSubscription } from '../../../modules/zuora/src/types/objects/subscription';
-import { zuoraDateFormat } from '../../../modules/zuora/src/utils/common';
 import type { ContributionTestAdditionalOptions } from '../../../modules/zuora/test/it-helpers/createGuardianSubscription';
 import { createContribution } from '../../../modules/zuora/test/it-helpers/createGuardianSubscription';
 import { doSwitch, preview } from '../src/contributionToSupporterPlus';
