+++ handlers/zuora-salesforce-link-remover/src/zuoraHttp.ts
@@ -1,7 +1,7 @@
 import { stageFromEnvironment } from '@modules/stage';
 import { ZuoraClient } from '@modules/zuora/zuoraClient';
-import { zuoraResponseSchema } from '../../../modules/zuora/src/types/httpResponse';
-import type { ZuoraResponse } from '../../../modules/zuora/src/types/httpResponse';
+import { zuoraResponseSchema } from '@modules/zuora/types';
+import type { ZuoraResponse } from '@modules/zuora/types';
 
 export async function updateBillingAccountInZuora(
 	zuoraBillingAccountId: string,
