+++ modules/zuora/src/auth/oAuthCredentials.ts
@@ -4,8 +4,8 @@ import {
 } from '@aws-sdk/client-secrets-manager';
 import { awsConfig } from '@modules/aws/config';
 import type { Stage } from '@modules/stage';
-import type { OAuthClientCredentials } from '../types/auth';
-import { oAuthClientCredentialsSchema } from '../types/auth';
+import type { OAuthClientCredentials } from '../types';
+import { oAuthClientCredentialsSchema } from '../types';
 
 export const getOAuthClientCredentials = async (
 	stage: Stage,
