+++ modules/zuora/src/billingPreview.ts
@@ -2,8 +2,8 @@ import { groupBy, sortBy, sumNumbers } from '@modules/arrayFunctions';
 import { getIfDefined } from '@modules/nullAndUndefined';
 import dayjs from 'dayjs';
 import type { Dayjs } from 'dayjs';
-import type { BillingPreview } from './types/objects/billingPreview';
-import { billingPreviewSchema } from './types/objects/billingPreview';
+import type { BillingPreview } from './types';
+import { billingPreviewSchema } from './types';
 import { zuoraDateFormat } from './utils/common';
 import type { ZuoraClient } from './zuoraClient';
