+++ modules/zuora/test/subscription.test.ts
@@ -5,12 +5,12 @@ import {
 	getSubscription,
 	getSubscriptionsByAccountNumber,
 } from '@modules/zuora/subscription';
-import type { ZuoraSubscription } from '../../../modules/zuora/src/types/objects/subscription';
-import { zuoraSubscriptionResponseSchema } from '../../../modules/zuora/src/types/objects/subscription';
-import { zuoraResponseSchema } from '../../../modules/zuora/src/types/httpResponse';
-import type { ZuoraResponse } from '../../../modules/zuora/src/types/httpResponse';
-import { zuoraSubscriptionsFromAccountSchema } from '../../../modules/zuora/src/types/objects/account';
-import type { ZuoraSubscriptionsFromAccountResponse } from '../../../modules/zuora/src/types/objects/account';
+import type { ZuoraSubscription } from '@modules/zuora/types';
+import { zuoraSubscriptionResponseSchema } from '@modules/zuora/types';
+import { zuoraResponseSchema } from '@modules/zuora/types';
+import type { ZuoraResponse } from '@modules/zuora/types';
+import { zuoraSubscriptionsFromAccountSchema } from '@modules/zuora/types';
+import type { ZuoraSubscriptionsFromAccountResponse } from '@modules/zuora/types';
 
 jest.mock('@modules/zuora/zuoraClient');
