+++ modules/zuora/test/zuoraSchemas.test.ts
@@ -1,5 +1,5 @@
-import { zuoraSubscriptionResponseSchema } from '../src/types/objects/subscription';
-import { zuoraSubscriptionsFromAccountSchema } from '../../../modules/zuora/src/types/objects/account';
+import { zuoraSubscriptionResponseSchema } from '@modules/zuora/types';
+import { zuoraSubscriptionsFromAccountSchema } from '@modules/zuora/types';
 import subscriptionJson from './fixtures/subscription.json';
 import subscriptionsFromAccountJson from './fixtures/subscriptions-from-account-number-response.json';
